Limit tests for Remove(K, out V) to netcoreapp
