+++ src/System.Collections/tests/Generic/Dictionary/Dictionary.Generic.Tests.cs
@@ -154,77 +154,5 @@ namespace System.Collections.Tests
         }
 
         #endregion
-
-        #region Remove(TKey)
-
-        [Theory]
-        [MemberData(nameof(ValidCollectionSizes))]
-        public void Dictionary_Generic_RemoveKey_ValidKeyNotContainedInDictionary(int count)
-        {
-            Dictionary<TKey, TValue> dictionary = (Dictionary<TKey, TValue>)GenericIDictionaryFactory(count);
-            TValue value;
-            TKey missingKey = GetNewKey(dictionary);
-
-            Assert.False(dictionary.Remove(missingKey, out value));
-            Assert.Equal(count, dictionary.Count);
-            Assert.Equal(default(TValue), value);
-        }
-
-        [Theory]
-        [MemberData(nameof(ValidCollectionSizes))]
-        public void Dictionary_Generic_RemoveKey_ValidKeyContainedInDictionary(int count)
-        {
-            Dictionary<TKey, TValue> dictionary = (Dictionary<TKey, TValue>)GenericIDictionaryFactory(count);
-            TKey missingKey = GetNewKey(dictionary);
-            TValue outValue;
-            TValue inValue = CreateTValue(count);
-
-            dictionary.Add(missingKey, inValue);
-            Assert.True(dictionary.Remove(missingKey, out outValue));
-            Assert.Equal(count, dictionary.Count);
-            Assert.Equal(inValue, outValue);
-            Assert.False(dictionary.TryGetValue(missingKey, out outValue));
-        }
-
-        [Theory]
-        [MemberData(nameof(ValidCollectionSizes))]
-        public void Dictionary_Generic_RemoveKey_DefaultKeyNotContainedInDictionary(int count)
-        {
-            Dictionary<TKey, TValue> dictionary = (Dictionary<TKey, TValue>)GenericIDictionaryFactory(count);
-            TValue outValue;
-
-            if (DefaultValueAllowed)
-            {
-                TKey missingKey = default(TKey);
-                while (dictionary.ContainsKey(missingKey))
-                    dictionary.Remove(missingKey);
-                Assert.False(dictionary.Remove(missingKey, out outValue));
-                Assert.Equal(default(TValue), outValue);
-            }
-            else
-            {
-                TValue initValue = CreateTValue(count);
-                outValue = initValue;
-                Assert.Throws<ArgumentNullException>(() => dictionary.Remove(default(TKey), out outValue));
-                Assert.Equal(initValue, outValue);
-            }
-        }
-
-        [Theory]
-        [MemberData(nameof(ValidCollectionSizes))]
-        public void Dictionary_Generic_RemoveKey_DefaultKeyContainedInDictionary(int count)
-        {
-            if (DefaultValueAllowed)
-            {
-                Dictionary<TKey, TValue> dictionary = (Dictionary<TKey, TValue>)(GenericIDictionaryFactory(count));
-                TKey missingKey = default(TKey);
-                TValue value;
-
-                dictionary.TryAdd(missingKey, default(TValue));
-                Assert.True(dictionary.Remove(missingKey, out value));
-            }
-        }
-
-        #endregion
     }
 }
